Implémenter système de coûts dynamiques et corriger affichage des niveaux
- Remplacer les coûts de déblocage arbitraires par un algorithme basé sur la difficulté
- Système d'écart progressif : 2 points (paliers 1-5), 1 point (6-15), 0 point (16+)
- Compatible avec système d'ordre personnalisé et mode statique
- Corriger le problème d'affichage vide des niveaux dans les paliers
- Synchronisation automatique des tiers et niveaux au démarrage
- Génération dynamique des tiers pour ordre personnalisé avec levelIds

+++ test/widgets/money_time_progress_screen_test.dart
@@ -262,8 +262,8 @@ void main() {
 
       for (final testCase in testCases) {
         await tester.pumpWidget(createTestWidget(
-          targetAds: testCase['targetAds']! as int,
-          selectedDuration: testCase['duration']! as int,
+          targetAds: testCase['targetAds']!,
+          selectedDuration: testCase['duration']!,
         ));
 
         // Verify correct information is displayed
